+++ src/Constants/IdStrings.php
@@ -47,6 +47,7 @@ class IdStrings
     public const INVOICE_FACTORING = 'ivf';
     public const INVOICE_GUARANTEED = 'ivg';
     public const INVOICE_SECURED = 'ivs';
+    public const PAYLATER_INVOICE = 'piv';
     public const PAYMENT_PAGE = 'ppg';
     public const PAYPAL = 'ppl';
     public const PIS = 'pis';
@@ -79,6 +80,7 @@ class IdStrings
         self::INVOICE_FACTORING,
         self::INVOICE_GUARANTEED,
         self::INVOICE_SECURED,
+        self::PAYLATER_INVOICE,
         self::PAYMENT_PAGE,
         self::PAYPAL,
         self::PIS,