+++ CHANGELOG.md
@@ -7,8 +7,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 *   Add Support for basket `v2` resource.
 ### Changed
-* Add support for payment state `create` which can occur when using payment pages.
-* Several minor improvements.
+*   Add support for payment state `create` which can occur when using payment pages.
+*   Examples:
+    *   Use `v2/basket` resource for secured payment methods and payment pages.
+    *   Remove broken image-URLs of payment page examples.
+*   Several minor improvements.
 
 ## [1.1.4.2](https://github.com/unzerdev/php-sdk/compare/1.1.4.1..1.1.4.2)
 ### Added