+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.6.0](https://github.com/unzerdev/php-sdk/compare/1.1.5.0..1.1.6.0)
+### Added
+* Add payment type Paylater Invoice.
+
 ## [1.1.5.0](https://github.com/unzerdev/php-sdk/compare/1.1.4.2..1.1.5.0)
 ### Added
 *   Add Support for basket `v2` resource.